backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.14.0
openai==0.19.0
django-cors-headers==3.13.0
gunicorn==20.1.0
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib